# apis/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moadim Server API",
    "description": "REST API for managing routines",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.5.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This server"
    }
  ],
  "paths": {
...
    "/restart": {
      "post": {
        "tags": [
          "crate::routes::restart"
        ],
        "summary": "`POST /restart` — stop this server and start a fresh instance.",
        "description": "The running server cannot rebind its own port, so it spawns a detached `moadim restart` helper\nthat stops it and starts a new process, mirroring the `moadim restart` CLI command and the\n`restart` MCP tool. Responds with the helper's PID before the restart completes.",
        "operationId": "restart",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestartResponse"
                }
              }
            }
          },
          "500": {
            "description": "could not spawn the restart helper"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RestartResponse": {
        "type": "object",
        "description": "Response body for `POST /restart` and the `restart` MCP tool.",
        "required": [
          "status",
          "helper_pid"
        ],
        "properties": {
          "helper_pid": {
            "type": "integer",
            "format": "int32",
            "description": "PID of the detached helper process performing the stop-old-then-start-new restart.",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "Acknowledgement status (always `\"restarting\"`)."
          }
        }
      },
...
    }
  }
}